Clean up folder naming and data type extension
