+++ Sources/FluentPostGIS/Geography/GeographicMultiLineString2D.swift
@@ -29,8 +29,3 @@ extension GeographicMultiLineString2D: GeometryConvertible, GeometryCollectable 
         self.geometry
     }
 }
-
-extension GeographicMultiLineString2D: PostGISDataType {
-    public static var dataType: DatabaseSchema
-        .DataType { PostGISDataTypeList.geographicMultiLineString }
-}
