+++ Sources/FluentPostGIS/Geometry/GeometricPoint2D.swift
@@ -31,7 +31,3 @@ extension GeometricPoint2D: GeometryConvertible, GeometryCollectable {
         self.geometry
     }
 }
-
-extension GeometricPoint2D: PostGISDataType {
-    public static var dataType: DatabaseSchema.DataType { PostGISDataTypeList.geometricPoint }
-}
